+++ react/OrderShipping.tsx
@@ -11,7 +11,7 @@ import {
   OrderForm as CheckoutOrderForm,
   Address as CheckoutAddress,
   DeliveryOption,
-  PickupOption
+  PickupOption,
 } from 'vtex.checkout-graphql'
 import EstimateShippingMutation from 'vtex.checkout-resources/MutationEstimateShipping'
 import SelectDeliveryOptionMutation from 'vtex.checkout-resources/MutationSelectDeliveryOption'
@@ -136,19 +136,23 @@ export const OrderShippingProvider: React.FC = ({ children }) => {
         },
       }))
 
-      enqueue(task, 'selectDeliveryOption').then(newOrderForm => {
-        if (queueStatusRef.current === QueueStatus.FULFILLED) {
-          setOrderForm(newOrderForm)
-        }
-      })
+      return enqueue(task, 'selectDeliveryOption').then(
+        newOrderForm => {
+          if (queueStatusRef.current === QueueStatus.FULFILLED) {
+            setOrderForm(newOrderForm)
+          }
 
-      return { success: true }
+          return { success: true, orderForm: newOrderForm }
+        },
+        () => ({ success: false })
+      )
     },
     [queueStatusRef, selectDeliveryOption, enqueue, setOrderForm]
   )
 
-  const handleSelectPickupOption = useCallback(async (pickupOptionId: string) => {
-    const task = async () => {
+  const handleSelectPickupOption = useCallback(
+    async (pickupOptionId: string) => {
+      const task = async () => {
         const {
           data: { selectPickupOption: updatedOrderForm },
         } = await selectPickupOption({
@@ -169,18 +173,23 @@ export const OrderShippingProvider: React.FC = ({ children }) => {
               ...pickupOption,
               isSelected: pickupOption?.id === pickupOptionId,
             })
-          )
+          ),
         },
       }))
 
-      enqueue(task, 'selectPickupOption').then(newOrderForm => {
-        if (queueStatusRef.current === QueueStatus.FULFILLED) {
-          setOrderForm(newOrderForm)
-        }
-      })
+      return enqueue(task, 'selectPickupOption').then(
+        newOrderForm => {
+          if (queueStatusRef.current === QueueStatus.FULFILLED) {
+            setOrderForm(newOrderForm)
+          }
 
-      return { success: true }
-  }, [queueStatusRef, selectPickupOption, enqueue, setOrderForm])
+          return { success: true, orderForm: newOrderForm }
+        },
+        () => ({ success: false })
+      )
+    },
+    [queueStatusRef, selectPickupOption, enqueue, setOrderForm]
+  )
 
   const handleSelectAddress = useCallback(
     async (address: CheckoutAddress) => {
@@ -237,7 +246,7 @@ export const OrderShippingProvider: React.FC = ({ children }) => {
       deliveryOptions,
       pickupOptions,
       handleSelectDeliveryOption,
-      handleSelectPickupOption
+      handleSelectPickupOption,
     ]
   )