+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Function `selectPickupOption` and `selectDeliveryOption` not returning orderForm on success case.
 
 ## [0.7.0] - 2020-12-02
 ### Added